refactor: weight to body namespace
