+++ unity-project/Assets/Scripts/Body/UI/Weight.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 40effd9bf11c5634f9198525125ceba6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
