+++ unity-project/Assets/Scripts/Body/Weight.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 20cad724eb4c4bd4699c42531e3c0904
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
